% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data Set}
\format{
## `fire damage`
A data frame with 15 observations and 2 variables:
\describe{
  \item{DISTANCE}{Distance to the fire.}
  \item{DAMAGE}{Damage caused by the fire}
}
}
\source{
Unknown
}
\usage{
fire
}
\description{
This data set contains information about fire incidents, including the distance of the fire
from a specific location and the resulting damage.
}
\keyword{datasets}
