% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{My Central Limit Theorem Uniform Function}
\usage{
mycltu(n, iter, a = 0, b = 10, x)
}
\arguments{
\item{n}{# an integer sample size}

\item{iter}{# an integer number of samples}

\item{a}{# lower bound for uniform}

\item{b}{# upper bound for uniform}

\item{x}{# a variable}
}
\value{
# none
}
\description{
My Central Limit Theorem Uniform Function
}
