% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycubic.R
\name{mycubic}
\alias{mycubic}
\title{Mycube function}
\usage{
mycubic(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A quantitative vector with cubed components
}
\description{
Mycube function
}
\examples{
mycubic(1:10)
}
