% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Maximum likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log likelihood function}

\item{x}{a quantitative variable}

\item{param}{a population parameter}

\item{...}{all other arguments}
}
\value{
a list of: i, the param at i, y at i, and the slope
}
\description{
My Maximum likelihood
}
