% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Myncurve function}
\usage{
myncurve(mu = 0, sigma = 1, a)
}
\arguments{
\item{mu}{A real number}

\item{sigma}{A real number}

\item{a}{A real number}
}
\value{
mu, sigma, and the area under the normal from -inf to a
}
\description{
Myncurve function
}
