% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{My n tickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats}

\item{gamma}{Acceptable overbooking probability}

\item{p}{Probability that an individual shows up}
}
\value{
list: nd (discrete), nc (continuous), N, gamma, and p
}
\description{
My n tickets function
}
\examples{
ntickets(400, 0.02, 0.95)
}
