% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My binomial distribution samples}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of plots}

\item{n}{Sample size}

\item{p}{Probability}
}
\value{
Plots
}
\description{
My binomial distribution samples
}
\examples{
mybin()
}
