% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My bootstrap 2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{A quantity}

\item{x}{A quantitative vector}

\item{fun}{A function name}

\item{alpha}{A significance value}

\item{cx}{CX}

\item{...}{All other arguments}
}
\value{
A graph of samples and a list of values
}
\description{
My bootstrap 2 function
}
