% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Optimal Ticket Sales for Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{An integer. The number of seats available.}

\item{gamma}{A numeric value between 0 and 1. The acceptable overbooking probability (the probability
that the number of ticket holders exceeding the available seats is within acceptable limits).}

\item{p}{A numeric value between 0 and 1. The probability that an individual ticket holder shows up.}
}
\value{
A list containing:
\item{nd}{The optimal number of tickets to sell using the discrete method.}
\item{nc}{The optimal number of tickets to sell using the normal approximation method.}
\item{N}{The number of seats available.}
\item{gamma}{The overbooking probability.}
\item{p}{The probability of a ticket holder showing up.}

The function also produces two plots:
- The first plot shows the objective function for the discrete calculation (blue curve) with a vertical
  red line indicating the optimal number of tickets to sell.
- The second plot shows the objective function for the normal approximation (black curve) with a vertical
  blue line indicating the optimal number of tickets to sell using the normal approximation.
}
\description{
This function calculates the optimal number of tickets to sell for an airline, taking into account
both discrete and normal approximation methods, based on a given overbooking probability (\code{gamma})
and the probability that a ticket holder shows up (\code{p}). It aims to minimize the probability of overbooking.
}
\examples{
ntickets(400, 0.02, 0.95)
# The function will calculate and display the optimal number of tickets
# to sell for an airline with 400 seats, 2\% overbooking probability, and 95\% show-up rate.

}
